% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDMA.R
\docType{data}
\name{shiny.maxRequestSize}
\alias{shiny.maxRequestSize}
\title{rstudio viewer에서 downloadhandler 사용 불가
save 후 데이터 다운로드
save 하면 데이터 수정이 될까 확인하기
function 더 추가하기}
\format{An object of class \code{numeric} of length 1.}
\usage{
shiny.maxRequestSize
}
\arguments{
\item{RDMA}{}
}
\description{
rstudio viewer에서 downloadhandler 사용 불가
save 후 데이터 다운로드
save 하면 데이터 수정이 될까 확인하기
function 더 추가하기
}
\keyword{datasets}
